Update to prioritize MP3 files
